rril: Ignore response when ignore flag set
BZ: 143718

Upon XREG URC, "Unidentified response" log is printed
which is confusing.

XREG URCs will be reported on changes in PS registration
state, band, act, lac or cid. If the XREG URC is received
only due to band change, response is set as unidentified.
This results in log "Unidentified response".

When the XREG URC is received only for band change, set
the response as unsolicited and also set the ignored flag.
Unsolicited flag is to confirm that it is a complete response
and the ignored flag is for print the response which
is ignored and proceed with processing next response.

Change-Id: I86eb0df75c4e53f28ba80018f13983ea5bc0484a
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/137327
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>