rril: Add "NO CARRIER" parsing in silo_common
BZ: 144447

The RRIL can receive a "NO CARRIER" URC on the STK channel
after a SET_UP_CALL proactive command. As this URC is not
handled by the silo_sim, we can have error with the parsing
of the URC coming after this "NO CARRIER".

This patch fixes the issue by adding parsing of the
"NO CARRIER" URC in the silo_common.

Change-Id: Ia00fc71262497fb38d2f9e73ca9ea6edd5f8a618
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/137859
Reviewed-by: Riviere, NathanX <nathanx.riviere@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>