rril: Enable signal strength reporting based on config flag
BZ: 144903

Currently, XCSQ URC is enabled by default.

With this patch, XCSQ URC will be enabled based on
configuration flag EnableSignalStrengthReporting.
If the EnableSignalStrengthReporting is set in
repository.txt, then the XCSQ URC will be used to
report signal strength changes.
If the EnableSignalStrengthReporting is not set or
set to 0 in repository.txt, then XCSQ URC will not be
used to report signal strength changes. In this case,
framework will poll the signal strength every 20seconds
in screen on state.

Change-Id: I03db7cf817cd3c278e3a5fede069007e5e9a20b7
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/138225
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>