rril: review sending of AT+CCLK
BZ: 145228

The AT+CCLK command is prepared a long time in advance during
RRIL init before to be sent to the modem. The bufferized AT
command is sent after the modem initialization. Due to this
we have a delta of 15 seconds between AP and BP logs.

This patch updates the RRIL code to create the AT+CCLK command
just before to send it to avoid the delta.

Change-Id: I48317d4e87a323ad7ffc89130b11221148401ca9
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/139399
Reviewed-by: Roulliere, Pierre <pierre.roulliere@intel.com>
Reviewed-by: Bedouet, Pascal <pascal.bedouet@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Haouati, Nizar <nizar.haouati@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>