rril: Enable +XCESQI to include LTE signal strength reporting
BZ: 125416

When the device is registered to LTE, correct LTE signal
strength is not shown.

When the device is registered to LTE, rapid ril fills
only the GSM signal strength structure based on values
reported by XCSQ. LTE signal strength information expected
by telephony framework is reported as part of XCESQI URC
and this URC is not yet enabled.

As XCESQI URC reports signal strength for the current serving
cell, enable XCESQI to report signal strength from modem 7160
and onwards.

Change-Id: I135fa8bd8f2f06fc10c1fd9b77d54d68997dfd8c
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>