rril: Improve handling of errors encountered while parsing URCs
BZ: 145903

If the parsing of an URC  fails, channel is blocked forever.
Due to this subsequent URCs are not parsed at all resulting
in wrong information(e.g: registration information) shown to
the user.

If parsing of an URC succeeds/fails, update the response pointer
correctly. Updating of the response pointer is now moved to a
centralized place.

Change-Id: I2a505557835f59abb7272fc08ad8efaf93c3db02
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/139162
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Tested-by: Haouati, Nizar <nizar.haouati@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>