rril: read the EnableCellInfo property only once
BZ: 136604

In the current implementation, the EnableCellInfo is read
each time that we need it. This configuration variable is
a static property located in the repository.txt file. So
we can read it only once and store the value internaly.

This patch improves the code by reading the configuration
property only once and storing it in the CTE.

Change-Id: I0c7343be13f23396ecbe9a0349e46a8aed936e71
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/139393
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Roulliere, Pierre <pierre.roulliere@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Haouati, Nizar <nizar.haouati@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>