rril: Allow RIL_UNSOL_OEM_HOOK_RAW in radio request pending
BZ: 146809

When the user receives an incoming call just before
airplane mode is activated, incoming call screen is shown
even when the call disconneced due to airplane mode
activation.

RIL_UNSOL_OEM_HOOK_RAW is not notified to framework when
the radio request(ON/OFF) is pending. Due to this, call
is still active on the android side whereas call is
disconnected on the modem side.

Allow RIL_UNSOL_OEM_HOOK_RAW notifications to be sent
to framework even when the platform shutdown is requested
or even when the radio request is pending.

Change-Id: I8058ff7bf74b8d6eac6173e509fe88aab98d0671
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/140033
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Tested-by: Haouati, Nizar <nizar.haouati@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>