rapid_ril: DNS parsing issue when mms is sent
BZ: 146883

As per the logs, when MMS context is established, it was seen
that XDNS? return 2 sets of DNSs, one for internet and another
for MMS. While parsing, when the address is 0.0.0.0 for the
second set, the first sets's DNS value is used, as the buffer
used for parsing was not reset.

This patch declares the buffers inside the AT+XDNS? response
parsing while loop.

Change-Id: I411892a0532eecd67618aa2a81cb41ce1ecda9fd
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/140349
Reviewed-by: buildbot <buildbot@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: Jin, AlbertX <albertx.jin@intel.com>
Tested-by: Jin, AlbertX <albertx.jin@intel.com>