rril: Fix CBS reception enabled even when deactivated
BZ: 146857

CBS reception is enabled even when the user has disabled the
CBS reception.

When the configuration list is empty, data provided for
RIL_REQUEST_GSM_SET_BROADCAST_SMS_CONFIG can be of length 0.
If the data length is 0, rapid ril considers it as an error.
Due to this, disabling of cell broadcast reception is not
informed to modem.

When the RIL_REQUEST_GSM_SET_BROADCAST_SMS_CONFIG is received
with data length of 0, consider it as valid case.

Change-Id: I6296ed5e13ac5949094a018f3fea6909d90579cc
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>