rril: fix time-out of AT+CGAUTO=1 command
BZ: 146254

As CGAUTO command was sent with NONE request ID, it was using
the default time-out of 10 seconds. This lead to APIMRs as the
modem may start (or wait for the end of) a PS attach procedure
before replying to this command.

The patch aligns the time-out for this command with the one of
the SETNETWORKSELECTIONMANUAL / SETNETWORKSELECTIONAUTOMATIC
requests (300 seconds).

Change-Id: Ie2bc59df4814087fbaddf5958af3b8568ecd63d0
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>