rril: Fix no data icon shown after MT call
BZ: 149776

After few seconds of MT call disconnect, data icon is
not shown even when the browsing works.

Upon CGEV: NW PDN DEACT 1, context is not cleaned up
properly resulting in context ID 1 not getting used for
the subsequent data call activation. So, when the device
registers back to LTE after MT call disconnect, data
registration state is completed as not registered because
the default context is not active.

Upon deactivation not initiated by user, forcefully clean up
the context.

Change-Id: I7b0a8fdb68b674a43ad6dfea34ea28830db87166
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>