rril: Fix no data seen after data stall
BZ: 145342

After data stall, call to establish a data call
results in CME ERROR.

When the data stall is seen after a switch from GSM only
to LTE/WCDMA/GSM, framework deactivates the data and tries
to reestablish the data. Request to reestablish the data
always results in CME ERROR: 148. As the default PDN context
is active for the same apn, it is enough to enable the routing
for default PDN. Since default PDN Apn information is not
updated always, rapid ril instead of enabling the routing for
default PDN tries to establish a 2nd PDP context.

Following fixes are done to address the issue
- Since the APN information is updated in the CGCONTRDP parsing
function, fix the issue by querying the context parameters
using AT+CGCONTRDP=<cid> command.
- If the deactivate request is received with PDP_RESET reason,
always send the deactivation request(AT+CGACT=0,<cid>) to modem.

Change-Id: I3f4ec8ab6e00cf962b16ff260dcab835162e5097
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>