rril: dont disable routing on default pdn deactivation
BZ: 152428

Upon user deactivating the data call, rapid ril sends
AT+XDATACHANNEL=0,1 to disable the routing for default
PDN. Due to this data is not working after the device
detaches and reattaches againt to the network.

With this patch, no AT command will be sent to modem
on user deactivating the default PDN but the interface
will be brought down on Android side. Also, don't send
any command to activate default PDN if the default PDN
is already active.

Note: This is an intermediate solution until a cleaner
way to disable data without detaching from network is
agreed.

Change-Id: Icb1489c4ec86af5c05caffe022713bdfe02569fd
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>