rril: rework network registration
BZ: 164575

Currently, network registration request is sent
to modem on RIL_REQUEST_SET_NETWORK_SELECTION_AUTOMATIC
and RIL_REQUEST_SET_NETWORK_SELECTION_MANUAL ril
requests. This requests are sent even before the
initial attach apn is set. Due to this, default
PDN is established with empty APNs which is not
correct as per the 3GPP specifications.

With this patch, following changes are done to address
the issue to avoid doing major changes in the android
telephony framework:
- Send the network registration request only
if the initial attach apn is set. If initial
attach is not set, then the network registration
request will be sent to modem only after the initial
attach apn is set.
- Upon set initial attach apn request, deregister from
the network, set the initial attach apn and register to
the network again.
- Reset the initial attach apn parameters only when
completing the RIL_REQUEST_GET_SIM_STATUS request
and if sim app state is unknown. This is done as during
fast flight mode activation and deactivation, sim status
remains ready resulting in SetInitialAttachApn not
sent by framework.

Change-Id: Iffe98bec0dc52c20ff70c0ac3bf4e28b6942e2a4
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>