rril: avoid long DTMF tone during call operation
BZ: 122114

A long DTMF tone can be heard on the remote side during
call swapping. The steps to reproduce the issue are the
following:

  1. Have two calls ongoing (one on hold)
  2. Randomly enter DTMF digits
  3. Swap the calls before the end of DTMF tones
  4. Repeat step 2 and 3

If a DTMF is ongoing when the RRIL receives the swap
request (or any other call operation), it stops the DTMF
before sending the +CHLD command. This last command is
sent whitout any waiting of the end of the DTMF stop
operation. So, if the start DTMF is not yet completed by
the modem side, we can have the +CHLD command sent before
the stop DTMF. This is the reason why a long DTMF tone can
be heard.

This patch fixes the issue by adding an event to wait for
the end of the stop DTMF operation. The call operation
(+CHLD) is sent only when the DTMF stop operation is fully
completed.

Change-Id: Ib8dabe4d0cc30df073028e7bc63563ddae7e7cf7
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>