rril: Fix issue in handling network initiated context
BZ: 153517

As part of modem initialisation, rapid ril sends command
AT+CGAUTO=4. This configures the modem to reject all
network initiated PDP contexts. Due to this, failure
seen with GCF certification.

With this patch, following changes done to fix the issue:
- Configure the modem to Turn off automatic response for
Packet Domain only.
- Upon +CGEV: NW ACT <p_cid>,<cid>,<event_type>, with
<event_type> set to 1, accept the network initiated context
activation if the device is registered to EPS or reject the
context activation if the device is not EPS registered.

Change-Id: I2914e99b95a7e8de9984e1e58b446c701143f97a
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>