rril: reset the RadioStateChangedEvent before AT+CFUN sending
BZ: 162360

In some cases, we can have the PostRadioPower function called
before the waiting of the RadioStateChanged event. If we the
event is reseted just before the wait, the trigger of this
event done in the PostRadioPower function will not be catched.

This patch fix the issue by reseting the event before the
sending of the AT+CFUN (so before any call to PostRadioPower)
instead of just before the wait.

Change-Id: I2ffc2a84258bd79ad44ccfc8fe20a2b26b71a59d
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>