Fix issue in usage of ConcatenateStringNullTerminate
BZ: 163369

In many places, ConcatenateStringNullTerminate is
called with remaining length of the output buffer
as the 2nd argument whereas the function expects
the maximum size of the output buffer as the 2nd
argument.

Fix the issue by passing maximum size of the output
buffer as 2nd argument to ConcatenateStringNullTerminate
function. Also, with this patch an issue seen in
ConcatenateStringNullTerminate function is fixed.

Change-Id: I95d5720de4f064e495fad946612c9967e48483f0
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>