Fix issue seen when mms and default apn is same
BZ: 161119

When the access point name is same for both MMS
and default connection is same, after sensing MMS
browsing is not possible anymore.

Since the APN name is same for both MMS and default
connection, already activated default PDN is used
instead of establishing a new connection. When the
framework deactivates the MMS connection after sending
of MMS, rapid ril brings down the interface and also
framework will cleanup the routing table. Due to this,
browsing won't be possible.

If the SETUP_DATA_CALL is requested for default data
profile and if the device is EPS registered, then use
the default PDN connection if the APN matches.

Change-Id: Id60e36d96ee4178ccac4809e9b677b5d2f282e9f
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>