rril: handle steering of roaming refresh type
BZ: 153296

The steering of roaming refresh is handled by the modem
side. For the AP side, this refresh is simply a SIM file
update on EF_OPLMNwACT.

This patch updates the RRIL code to handle this refresh
as a SIM file update instead of a SIM reset. No need to
specify that the EF_OPLMNwACT is updated as this file is
not used by Android.

Change-Id: I49f7ffc052ad8fcd530e2d0036584ebbe7f883a8
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>