rril: avoid usage of CInitializer after free
BZ: 164217

The instance of CInitializer is deleted when the init
fails in CSystemManager::InitializeSystem() and is accessed
later in the destructor of CSystemManager.

This patch fixes the issue by setting to NULL the pointer
of the CInitializer instance when this one is deleted and
by adding a check in the destructor of CSystemManager before
to access it.

Change-Id: Ia8defe7d989e2c0469e3beb2dc2ac1b805956bb3
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>