rril: fix unefficient snprintf usage
BZ: 164899

Due to bad interpretation of snprintf man page, some code in the
Rapid RIL manually null terminated the string (and wrote using a
buffer length truncated by one).

As snprintf always null terminates the string, this is useless
and unefficient (as one byte of the destination buffer is
wasted).

Change-Id: I8d41fa4bb64d82744d37a37f09310c02697b3445
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>