rril: Class S implementation, dynamic management of UICC profiles
BZ: 156976

Class "S" implementation: The modem implements a new USIMAP
component that need to be fed with a TE profile and a MT profile
to cleanly separate the proactive commands handled by AP from
proactive commands handled by the modem. That follows the new
Rel.10 Class "s", CAT over the modem interface, feature
described in ETSI TS 102223 V10.6.0, specialized in
3GPP TS 31.111 V10.10.0.

That means we have to handle in RRIL a new profile
initialization phase that will configure the profiles
if necessary before the USAT (AP) activation.

Change-Id: I066bbb371dfec9539f22eb7d2abf618727c0046c
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>