Deactivate all data calls only if not LTE registered
BZ: 172703

Conformance test 36.523-16.1.1.3 fails as there is an
unexpected RRC connection when searching for networks.

Inorder for the manual network scan to succeed all the
time, deactivation of all data calls has been done.

With this patch, deactivation of all data calls is done
only if the device is not registered to LTE.

Change-Id: Ib60ddd66aee5d05ef09d0cc80155c718e6cbd05a
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>