rril: Fix issue in parsing XCESQ <rssnr> parameter
BZ: 172707

Parsing of <rssnr> fails resulting in signal strength
not updated to the framework correctly.

As per C-AT specification, possible values for <rssnr>
include negative values. As the rapid ril is not able
to parse the negative value, XCESQ response is considered
as error and the response is ignored.

Fix the issue by parsing the negative values reported
for <rssnr> correctly.

Change-Id: I3691462a5df8336775bdc92822727c72228c58af
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>