Fix data icon not shown on some cases
BZ: 171966

Data Icon not displayed but browsing works in the
following case:
- Device is in airplane mode
- Screen state is off (done by pressing power button)
- After modem is completely powered down in flight mode,
user brings up the screen and deactivates the airplane
mode.
- After deactivation of the airplane mode, device is
registered to network, browsing works but no data icon
is shown.

Rapid ril rejects the SCREEN_STATE request when the
modem is not available. Due to this, only registration
state reporting is enabled. So, when the device registers
to network after airplane mode deactivation, rapid ril
based on screen state has enabled only registration state
reporting. This results in browsing working but no data icon
shown as access technology reporting is not enabled due to
wrong screen state maintained in the rapid ril.

Fix the issue by caching the screen state also in
modem not available state.

Change-Id: I0b9692f10b57a4f4a09ccef4aa531aba3d67a1d6
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>