Fix mms issue seen when mms and default apn is same
BZ: 177250

With CMCC HK operator, mms retreival is not working if wifi
is connected. This issue is seen when the device is registered
to LTE and the apn name for default, mms connection is same.

Context activation request for MMS connection is rejected
by network. Due to this, MMS retreival doesn't work.

Following changes are done to fix the issue:
- Reference count and routing information is maintained for each
active connection.
- On SETUP_DATA_CALL request with same apn as default PDN
  - if the default PDN is active, routing is enabled and
    reference count is > 0, complete the request with default
    PDN connection information.
  - If the default PDN is active but routing not enabled, enable
    the routing and complete the request with default PDN connection
    information.
- Upon deactivation request, remove the interface on force cleanup
or reference count is 0.
- Change the default PDN state to E_DATA_STATE_ACTIVE after
reading the default PDN context parameters and set the routing
enabled to false.
- RoutingEnabled is set to TRUE only after the routing is enabled
by sending XDATACHANNEL commands. Note: For 6260 modem, this is
done after the activation of context.

Change-Id: Id485bc78df98ff0b65cff468b952746015571770
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>