rril: keep usbX interface when data context up at modem side
BZ: 160864

Due to the inability of tearing down the default PDN when LTE
camped, in some cases of data deactivation, nothing is done at
modem side (but usbX networking interface was brought down).

This meant that if packets arrived at modem side after this
'pseudo' tear down, it would wake-up the AP to send them but as
usbX was down, nothing was present at AP side to flush them
leading to a severe current drain impact.

Fix (waiting for support for a new AT command at modem side) is
to tear down usbX interface and then bring it back up (without
any IP address associated to it) so as to flush modem packets.

Change-Id: Iab306fb2775949c248f3b5744f223414573dfc14
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>