On CSFB, fix wrong access technology shown after call ended
BZ: 177253

Upon incoming call, screen is switched on to notify the
incoming call to user. When the screen is switched on,
registration status is queried from the modem. This queried
registration status is not updated in the m_PreviousXREGInfo.
When the call is ended XREG URC is received with LTE information
but it is ignored. As the response to registration status query
is not updated in m_PreviousXREGInfo, reported registration
state is ignored even when there is a change in ACT.

Fix the issue by comparing the received registration status
with the previous registration status cached in the CTE class.

Change-Id: I92cebfae7c080dbed56992f4b9ab40d4c3afce42
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>