RRIL: rework interface for COPS=0
BZ: 82954

We need to distinguish cases where COPS=0 response
is not received due to lack of coverage vs cases where modem is dead.
The solution is to tag COPS=0 command as special case:
- if timeout happens then we PING modem to check if it is still alive.
- if PING is OK: re-launch a timeout
- do this until COPS is finally answered.

Because DLC2 is used for other commands also, blocking
indefinitely this channel is not acceptable; for this reason
the COPS=x commands will be sent on DLC22.

Change-Id: I8db32174b965e9584dd89ed42ff5e0b64e5059c1
Signed-off-by: Goea, Andrei DanielX <andrei.danielx.goea@intel.com>