rril: do not call AT+XSATK for te_xmm7260
BZ:178642

CSilo_SIM::GetURCInitString() still calls directly
+XSATK which is from the old USAT interface.

The call is done for the scenario when the platform
does not use USAT : in previous interface it was
necessary to do this call +XSATK=1,0 that allows
proactive command to be fetched, but no proactive
command is sent to the AP side.

This at command is not used for 7260 modems.

Change-Id: If7797698ea0174a596d203081cfbb9d7fd0e1bb7
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>