rril: Replace usage of +XRAT by +XACT (except for DSDS)
BZ: 175819

As Xmm6360 is a Rel.9 modem, WPRD recommends to use +XACT
and not +XRAT which is not officially supported on Xmm6360.

The only exception if for the feature DSDS swap RAT mode
between the 2 stacks (+XRAT=8) that do not have an
equivalent in +XACT, where AT+XRAT=8 must be kept.

Change-Id: Iabb699510089ae64a888513c073c1d5cc27c0cd7
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>