Fix issue in ucs2_to_utf8 conversion which leads to modem CD
BZ: 184751

When the sim application label is UCS2 coded, application
label overwrites the session id. This results in logical
channel access command CRLA sent with invalid session id.
Modem core dumps when negative session id is sent for
logical channel access.

With this patch, following changes are done to fix the
issue:
- Don't allow logical channel access if the session id
is <= 0.
- Convert the UCS2 to UTF8 only if the value is valid.

Change-Id: I47ec69a0365c07c672027cb33c783baf46cb7fe5
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>