rapid_ril: Correct handling of DATA CALL LIST
BZ: 187715

The DATA CALL LIST query from android should return
current active data connections, but in case of LTE
when a data connection is disconnected due to wifi
activity, the information returned is not correct.

In case of LTE the default PDN is never disconnected
but the interface is brought down on pdp deactivation
requests from Android framework. The ref count for the
channel data will indicate whether the context is being
used or not. This patch modifies the handling of
DATA_CALL_LIST by checking the ref count, ensuring the
information of data which is active and used is only
returned.

Change-Id: Ie9f3736f1042b4c7347cc29bb2c7ee83b2ead779
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>