Add support for registration state and band info reporting
BZ: 184997

For LTE coex handling, registration state and band
information needs to be reported to android framework.

Following changes are done to support this:
- New oem hook is added to activate or deactivate
registration state and band information reporting.
- Upon XREG URC, if the previous registration state
or the band information is not equal to the received
state or band information, registration state and
band information is reported to the framework.

Change-Id: I16e71bda3518cfa4b2a7811c51be1703ffc0e147
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>