rril: fix issue in reporting registration reject cause
BZ: 206926

Registration reject cause is not parsed even when it
is reported via voice registration state. Due to this,
registration reject cause is not reported to telephony
framework and eventually not presented to the user.
Also, 3GPP specific reject cause is getting overwritten by
the manufacture specific reject cause resulting in test case
failure.

Following changes are done to fix the issue:
- Parse the registration reject cause when the registration
state is either denied or emergency only.
- In case of manufacturer reject cause and on product config
property set to "att", use the previous reject cause stored in
the registration state cache.

Change-Id: I3ca85ea20d2e6eab03e2dcc835ec81c449d48460
Signed-off-by: georgiaw <georgia.wang@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Nicolas LOUIS <nicolasx.louis@intel.com>