rril: Modify handling of call number presentation
BZ: 199983

Currently the Android framework does not handle correctly the
case where the calling party does not send the distant phone number
or explicit CLIR information.

In fact if no number is sent the CLI_validity is set to
"restricted" by default when it should be set to "not available",
in this case the CNAP name could still be printed when it is not
currently because the CLI_validity is prior to CNI_validity.

CLI_validity is retrieved by enabling +CLIP URC for XMM7160
CLI_validity is retrieved from +CLCC URC for XMM7260

Change-Id: If56d94f214942d4cf9ace33175feccd8ac46a5d4
Signed-off-by: Morgane BUTSCHER <morganex.butscher@intel.com>
Signed-off-by: Petru MURARU <petrux.muraru@intel.com>