Fix device not registering on 1st boot after flash
BZ: 199829

On 1st boot after flashing, device doesn't register
to the network.

On 1st boot after flashing, rapid ril downloads
the terminal profile and resets the sim. As a result
of sim reset, XSIM: 0 is received which inturn resets
the cached network selection mode. So, when the modem
is ready for attach, network selection is not restored
as the network selection mode maintained in the rapid
ril is invalid.

Fix the issue by resetting stored network selection mode
on modem events and on radio power on/off.

Change-Id: Iacdbf3f52f48952ee4df4ce0d842fad33cc53d77
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>