Fix out of service seen after waking up from deep sleep mode
BZ: 200350

Device stays in out of service even when the device is
attached to LTE network and also when there is an
active data connection on modem side.

When there is no data transferred during the inactivity timer
period, default PDN is deactivated on modem side but the
PS is still attached and there is also an another active PDP
connection. As the default PDN is not active, rapid ril reports
the registration state as not registered. Also, LAC and CID
are not reported to framework when device registers to
network in screen off state.

Following changes are done to address the issue:
- Fix the issue by reporting data registration status as received
from modem when there is an active PDP connection.
- In screen off state, always query the registration status
from modem. This is achieved by setting m_bPSStatusCached and
m_bCSStatusCached to FALSE if the screen state is off.
- As screen state is cached, allow RIL_REQUEST_SCREEN_STATE
request even when modem is not available.
- Upon data registration state ril request, query both EPS
and GPRS/UMTS registration status from modem.

Change-Id: I8bec6de5f4422bb63a35449821f276550b438bc5
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>