Fix out of service seen while performing invactivity timer test
BZ: 200357

Device stays in out of service even when the device is
attached to LTE network and also when there is an active
data connection on modem side.

When there is no data transferred during the inactivity timer
period, data connection is deactivated on modem side but the
PS is still attached and IMS is registered. As the default
PDN is not active, rapid ril reports the registration state
as not registered but didn't update the registration status when
the data connection is established again after some period.

Fix the issue by reporting registration status as received from
modem when the device is IMS registered or when there is an
active data connection.

Change-Id: Idea66fab7ca10abaad64c0bc783375bf85a28e94
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>