rril: request P-CSCF only for IMS data profile
BZ: 198974

In the current code, the P-CSCF request flag is set to
true when we are in IMS AP-centric mode whatever the
data profile type. This must be done only for the IMS
data profile.

This patch updates the code to set the P-CSCF request
flag to true only for the IMS data profile type and
when the IMS mode is AP-centric.

Change-Id: Ibfd56ad3ebafcf5e71d5260eddf06b7c17c52387
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>