rril: fix ims deactivation
BZ: 202698

At boot the ril send the AT command to activate ims
in the modem. If the user have manually deactivated ims
in the ims settings menu, the oemtelephony will trigger
the deactivation of ims once radio is on. This lead to
a situation where ims is activated and then deactivated
few seconds later. This patch remove this behavior,
the  activation or deactivation of ims is done directly
based on the ims activation state set in the ims menu
settings.

Change-Id: I5cd3b8861a047d5bb40539086cb65530ae49fd41
Signed-off-by: alain kouassu <alain.kouassu@intel.com>
Signed-off-by: Nicolas LOUIS <nicolasx.louis@intel.com>