rril: properly release modem in all cases of RADIO_POWER OFF
BZ: 207186

In cases where the modem is not in a stable state when the
RADIO_POWER message "OFF" is received from the framework, the
RIL did not release the radio at MMGR side (but still returned
OK to the framework). This lead to cases where the phone is in
airplane mode but the radio was still on.

Fix the issue by ensuring that the radio is properly released
by RRIL at MMGR level on any reception of a RADIO_POWER "OFF"
message.

Change-Id: I00d36ff88f74e51e6b22c419431c62e9e9f13be9
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>