Rapidril: rework on registration status report for CS
BZ: 202021

Android framework decides concurrent voice and data
support based on the CS registration network type. In telephony
framework, if the network type is either or above UMTS, it is
considered to not to support voice-data concurency. But there is
an issue with the network type values, where the value for GSM
will also be reported to support concurrency. See ril.h

Rapid ril always reports the PS registration network type as
the network type for CS. For example in 2G rapid ril reports both
PS and CS registration as EDGE, where as EDGE is a data ( or PS)
technology.

These two anomalies create the issue mentioned in the BZ.

The issue is fixed by modifying rapid ril to report CS and PS
registration truely based on the actual network type. While in
CFSB, the reporting is based on the +XREG, URC and in all other
cases the CS registration is modified to be based on +CREG.
The mapping of URCs types to android type is modified to take
into account the URC type and report CS and PS network types
accordingly.

Change-Id: I176d5184f033896f45468373ccf01e5eada23fe5
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>