fix issue in handling initial attach apn parameter changes
BZ: 204448

Changes in initial attach parameters is not taken into account
immediately.

When the initial attach parameters are changed by user,
request to update the initial attach parameters is sent to
modem but it is rejected as the default PDN is active.

If there is a change in initial attach apn parameters,
deregister first before sending the initial attach apn
parameters to modem. Reset stored initial attach apn
and network selection mode parameters if sim is removed.
- Also with this patch, apn is updated in CChannel_Data
on SETUP_DATA_CALL request from framework.
- On response to AT+CGCONTRDP=<cid> command, update
the apn in CChannel_data only if apn is not empty.

Change-Id: Id1e784c1856c115f1204157ea301c23307c5047a
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>