rril: fix CREG/CGREG/CEREG parsing issue
BZ: 206365

The distinction between URC and command response is done by
relying on the parameters count.
This is not correct as most of the parameters are optional, we
can have cases where the numbers of parameters is the same between
URC and response
A simpler way of doing it is to parse the first 2 parameters:
- if there is only 1 param and it's an integer then URC
- if the second parameter is not an integer then URC
- otherwise it's a response
Error cases during the parsing will be handled by the original
parse functions.

Change-Id: I88b000ff6cc925d2e95141c8aee7c0a6cf18e78d
Signed-off-by: Frederic Predon <frederic.predon@intel.com>