Fix issue in reporting updated tac and cid
BZ: 207699

Operator test fails as the cell broadcast message is
rejected as duplication even when the message is received
in different tracking area.

When there is a tracking area change, framework is expecting the
updated change to be reported as part of voice registration
status query. As rapid ril is not completing the voice
registration status with updated tracking information, message
is rejected as duplicate.

Fix the issue by copying the TAC/CID received as part of EPS
registration status to voice registration status LAC/CID.

Change-Id: I239f98534fd11d3157bb07d5161c7fc81b4a1ab7
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Riviere, Nathan <nathanx.riviere@intel.com>