report cell information received via XMCI without filtering
BZ: 203707

cell information received with invalid mcc and mnc is not
reported to framework. Due to this, no neighboring cell is
reported to framework.

As modem doesn't know the mcc and mnc of the neighboring cell
it will report unknown values for mcc and mnc. Filtering based
on mcc and mnc unknown values is resulting in neighboring cell
information not reported to android.

Fix the issue by not doing any filtering of XMCI response.

Change-Id: I79da11ed6b3bdfc8be729e9b5b3b20160b6d2af6
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Signed-off-by: Nicolas LOUIS <nicolasx.louis@intel.com>