Fix device not registering on 1st boot after blankphone flash
BZ: 208773

When modem is ready(XSIM: 7) for attach, SIM status is notified
to framework. After reading sim files, framework sends request
RIL_REQUEST_SET_INITIAL_ATTACH_APN to set the initial attach
apn. This request is rejected by rapid ril as the sim state has
changed from ready to attach to not present due to +XSIM: 0.
With 7260 modem, +XSIM: 0 is received as rapid ril resets the
SIM after read and update of terminal profile.

Fix the issue by notifying the sim app state as detected if
the uiccState is not active or if the sim event is not
SIM_READY_TO_ACTIVATE, STK_SERVICE_RUNNING and if the sim app
state is ready.

Change-Id: If7bf1e8620eb53eb28b88350188beda954ab5df9
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>